requirements.txt
Django==1.3
argparse==1.2.1
django-haystack==1.2.7
django-piston==0.2.3
psycopg2==2.4.5
pysolr==2.1.0-beta
python-ptrace==0.6.4
simplejson==2.6.1
wsgiref==0.1.2